# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Comment:
    properties:
      body:
        type: string
      created_at:
        type: string
      expense_id:
        type: integer
      id:
        type: integer
      settlement_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
...
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      push_enabled:
        description: |-
          PushEnabled is the user's own preference, checked before sending any
          push notification — separate from whether they have any subscribed
          devices (PushSubscription rows).
        type: boolean
      updated_at:
        type: string
    type: object
  handler.AddCommentRequest:
    properties:
      body:
        example: Can we split this differently?
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for SplitEasy app (Splitwise style expense sharing).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SplitEasy API
  version: "1.0"
paths:
...
  /comments/{id}:
    delete:
      description: Soft-deletes a comment. Only its author may delete it.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Delete a comment
      tags:
      - comments
...
  /expenses/{id}/comments:
    get:
      description: Returns every comment on an expense, oldest first. Any member of
        the expense's group may view them.
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: List an expense's comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Adds a comment to an expense. Any member of the expense's group
        may comment.
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handler.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Comment on an expense
      tags:
      - comments
...
  /settlements/{id}/comments:
    get:
      description: Returns every comment on a settlement, oldest first. Any member
        of the settlement's group may view them.
      parameters:
      - description: Settlement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: List a settlement's comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Adds a comment to a settlement. Any member of the settlement's
        group may comment.
      parameters:
      - description: Settlement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handler.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - JWT: []
      summary: Comment on a settlement
      tags:
      - comments
...
securityDefinitions:
  JWT:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"